style pages/mainmenu buttons
