+++ client/src/pages/order/index.js
@@ -1,0 +1,30 @@
+import React from 'react';
+import { makeStyles } from '@material-ui/core/styles';
+import Link from '@material-ui/core/Link';
+import Typography from '@material-ui/core/Typography';
+import { Grid } from '@material-ui/core';
+
+const useStyles = makeStyles((theme) => ({
+  root: {
+    '& > * + *': {
+      marginLeft: theme.spacing(2)
+    }
+  }
+}));
+
+function Order () {
+  const classes = useStyles();
+  return (
+    <Grid container>
+      <Grid item xs={2} />
+      <Grid item xs={8}>
+        <Typography className={classes.root}>
+          Start an order
+        </Typography>
+      </Grid>
+      <Grid item xs={2} />
+    </Grid>
+  );
+}
+
+export default Order;
